chore(cloudtrail): move to jest
