+++ packages/@aws-cdk/aws-cloudtrail/.gitignore
@@ -13,3 +13,5 @@ nyc.config.js
 .LAST_PACKAGE
 *.snk
 !.eslintrc.js
+
+!jest.config.js
