+++ packages/@aws-cdk/aws-cloudtrail/.npmignore
@@ -19,3 +19,5 @@ dist
 
 tsconfig.json
 .eslintrc.js
+
+jest.config.js
