+++ packages/@aws-cdk/aws-cloudtrail/jest.config.js
@@ -1,0 +1,10 @@
+const baseConfig = require('../../../tools/cdk-build-tools/config/jest.config');
+module.exports = {
+  ...baseConfig,
+  coverageThreshold: {
+    global: {
+      branches: 75,
+      statements: 80,      
+    }
+  }
+};
