+++ packages/@aws-cdk/aws-cloudtrail/package.json
@@ -47,7 +47,8 @@
     "compat": "cdk-compat"
   },
   "cdk-build": {
-    "cloudformation": "AWS::CloudTrail"
+    "cloudformation": "AWS::CloudTrail",
+    "jest": true
   },
   "keywords": [
     "aws",
@@ -63,13 +64,12 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@aws-cdk/assert": "0.0.0",
-    "@types/nodeunit": "^0.0.31",
     "aws-sdk": "^2.679.0",
     "cdk-build-tools": "0.0.0",
     "cdk-integ-tools": "0.0.0",
     "cfn2ts": "0.0.0",
     "colors": "^1.4.0",
-    "nodeunit": "^0.11.3",
+    "jest": "^25.5.4",
     "pkglint": "0.0.0"
   },
   "dependencies": {
